oauth2: support multiple redirect urls per client
Added support for multiple redirect urls per client. When client sends
authorization request to the oauth2 provider it may request and redirect_uri
which to be preferred for the callback redirect. If it's not specified then the first
one of the list is used.

In cases when the requested URL is fake the oauth2 provider will return
bad request with proper information message to indicate that redirectUrl
was broken.

Fixes #23