oauth2: protect auth requests from CSRF
Clients are now able to pass state which is dispatched back to the
callback url.

State value could be used by the client to maintain state between
the request and callback.  The authorization server includes
this value when redirecting the user-agent back to the client.
The parameter SHOULD be used for preventing cross-site request
forgery as described in https://tools.ietf.org/html/rfc6749#section-10.12

An example response is shown in:
https://tools.ietf.org/html/rfc6749#section-4.1.2