oauth2/integration: better library integration
The library had a strog dependency to the HttpServlet. All the endpoint
setups were configured in the OAuth2Servlet, so when we try to integrate
the oauth2 library in our code we should use this servlet. This makes
the integration a litle bit restrictive and cause troubles if would like
to use frameworks like SparkJava.
Now the configurations logic was moved out and separated from the need to
use directly the HttpServlet in your project