oauth2-server: improve client authorization API
The client authorization api is now improved by using kotlin sealed
classes for results which delegates bigger role on the client to ensure
whether the client is allowed to use the application or not.

As side change the URL composing is improved by insing Urlsome library
which ensures that urls are properly created and are url safe.